+++ sherpa-onnx/csrc/features.h
@@ -14,7 +14,7 @@
 namespace sherpa_onnx {
 
 struct FeatureExtractorConfig {
-  float sampling_rate = 16000;
+  int32_t sampling_rate = 16000;
   int32_t feature_dim = 80;
   int32_t max_feature_vectors = -1;
 
@@ -34,7 +34,7 @@ class FeatureExtractor {
      @param waveform Pointer to a 1-D array of size n
      @param n Number of entries in waveform
    */
-  void AcceptWaveform(float sampling_rate, const float *waveform, int32_t n);
+  void AcceptWaveform(int32_t sampling_rate, const float *waveform, int32_t n);
 
   /**
    * InputFinished() tells the class you won't be providing any
